Agregar infisical para administrar secretos

+++ src/bin/deleteDBData.ts
@@ -1,12 +1,19 @@
-import { prismaClient } from "../utils/lib/prisma-client";
+import { SecretsManager } from "../utils/classes/SecretsManager";
 
 const deleteDBData = async () => {
+    let prismaClient;
 
     try {
         console.time("Tiempo de ejecucion");
         console.log(
             "------------ INICIANDO ELIMINACION DENTRO DE DB -------------\n",
         );
+
+        await SecretsManager.getInstance().initialize();
+        prismaClient = (await import("../utils/lib/prisma-client")).prismaClient;
+        if (!prismaClient) {
+            throw new Error("Prisma Client no inicializado");
+        }
 
         await prismaClient.familiarScout.deleteMany({})
         await prismaClient.familiar.deleteMany({})
@@ -21,7 +28,9 @@ const deleteDBData = async () => {
     } catch (error) {
         console.log("Error en el script: ", (error as Error).message);
     } finally {
-        await prismaClient.$disconnect();
+        if (prismaClient) {
+            await prismaClient.$disconnect();
+        }
     }
 };
